LayoutTests/fast/js/toString-elision-trailing-comma-expected.txt

This test checks that toString() round-trip on a function that has a array with elision does not remove a comma.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ueuf1 is uf1
PASS ueuf2 is uf2
PASS successfullyParsed is true

TEST COMPLETE
